提交

+++ tp5/application/admin/controller/Index.php
@@ -3,6 +3,7 @@ namespace app\admin\controller;
 
 use think\Loader;
 use think\Db;
+use function GuzzleHttp\json_encode;
 
 class Index extends \app\common\controller
 {
@@ -58,11 +59,40 @@ class Index extends \app\common\controller
 
     public function total_day()
     {
-        
+
         $page = 1;
         $page_size = 20;
-        $page_count = count(Db::table('cj_ad_click_log')->select());
-        $loglist = Db::query("select count(id) as idn,SUM(activate_state) as asn , CONCAT(YEAR(create_time),'-',MONTH( create_time),'-',DAY(create_time)) AS data_time from cj_ad_click_log GROUP BY data_time DESC");
+        $where = " where 1=1  ";
+        if($this->request->param("page")){
+            $page = $this->request->param("page");
+        }
+        
+        if (\json_decode(session('userinfo'),true)['user_id'] != 1 )
+        {
+            $where .= ' and b.add_channel_id='.\json_decode(session('userinfo'),true)['user_id'];   
+        }else{
+            if($this->request->param('add_channel_id')){
+                $where .= ' and b.add_channel_id='.$this->request->param('add_channel_id');     
+            }
+        }   
+        if($this->request->param('app_id')){
+            $where .= ' and a.app_id='.$this->request->param('app_id');
+        }
+        if($this->request->param('ad_name')){
+            $where .= ' and b.ad_name="'.$this->request->param('ad_name').'"';
+        }
+        if($this->request->param('app_name')){
+            $where .= ' and b.app_name="'.$this->request->param('app_name').'"';
+        }
+        if($this->request->param('create_time')){
+            $where .= ' and TO_DAYS(a.create_time)= TO_DAYS("'.$this->request->param('create_time').'")';
+        }
+        if($this->request->param('activate_time')){
+            $where .= ' and TO_DAYS(a.activate_time)= TO_DAYS("'.$this->request->param('activate_time').'")';
+        }
+        $limit = " limit ".($page-1)*$page_size.",".$page_size;
+        $page_count = count(Db::query("select count(a.id) as idn,SUM(a.activate_state) as asn ,SUM(a.is_call_next) as asn_y ,CONCAT(DATE_FORMAT(a.create_time,'%Y') ,'-',DATE_FORMAT(a.create_time,'%m'),'-',DATE_FORMAT(a.create_time,'%d')) AS data_time from cj_ad_click_log as a left join cj_ad_app as b on a.app_id=b.app_id $where GROUP BY data_time DESC order by data_time DESC"));
+        $loglist = Db::query("select count(a.id) as idn,SUM(a.activate_state) as asn ,SUM(a.is_call_next) as asn_y ,CONCAT(DATE_FORMAT(a.create_time,'%Y') ,'-',DATE_FORMAT(a.create_time,'%m'),'-',DATE_FORMAT(a.create_time,'%d')) AS data_time from cj_ad_click_log as a left join cj_ad_app as b on a.app_id=b.app_id $where GROUP BY data_time DESC order by data_time DESC $limit");
         // $applist = Db::table('cj_ad_click_log')
         //     ->alias('a')
         //     ->join('cj_ad_app b','a.app_id = b.app_id','LEFT')
@@ -109,33 +139,49 @@ class Index extends \app\common\controller
     {
         $page = 1;
         $page_size = 20;
-        $page_count = count(Db::table('cj_ad_click_log')->select());
+        if (\json_decode(session('userinfo'),true)['user_id'] != 1 ){
+            $page_count = count(Db::table('cj_ad_click_log')->select());
+        }else{
+            $page_count = count(Db::table('cj_ad_click_log')->select());
+        } 
         $applist = Db::table('cj_ad_click_log')
+            ->field('b.ad_name,b.app_name,b.ad_platform_id,b.add_channel_id,b.app_id,b.in_price,b.out_price,a.user_idfa,a.create_time,a.activate_state,a.activate_time,a.id')
             ->alias('a')
             ->join('cj_ad_app b','a.app_id = b.app_id','LEFT')
             ->order('a.id', 'desc');
-        if($this->request->param()){  
-            // if($this->request->get("name")){
-            //     $name = $this->request->get("name");
-            //     $applist = $applist->where('app_name','like',"%".$name."%");
-            // }
-            // // if($this->request->get("chaname")){
-            // //     $chaname = $this->request->get("chaname");
-            // //     $applist = $applist->where('ad_name','like',"%".$des."%");
-            // // }
-            // if($this->request->get("des")){
-            //     $des = $this->request->get("des");
-            //     $applist = $applist->where('ad_name','like',"%".$des."%");
-            // }
-            // if($this->request->get("id")){
-            //     $id = $this->request->get("id");
-            //     $applist = $applist->where("app_id",$id);
-            // }
+     
+            if($this->request->param('create_time')){
+                $applist = $applist->where("TO_DAYS(a.create_time)",'$TO_DAYS("'.$this->request->param('create_time').'")');
+            }
+            if($this->request->param('activate_time')){
+                $applist = $applist->where("TO_DAYS(a.activate_time)",'$TO_DAYS("'.$this->request->param('activate_time').'")');
+            }
+            if(\json_decode(session('userinfo'),true)['user_id'] != 1){
+                $add_channel_id = \json_decode(session('userinfo'),true)['user_id'];
+                $applist = $applist->where("b.add_channel_id",$add_channel_id);
+            }else{
+                if($this->request->param("add_channel_id")){
+                    $add_channel_id = $this->request->param("add_channel_id");
+                    $applist = $applist->where("b.add_channel_id",$add_channel_id);      
+                }
+            }
+            if($this->request->param("app_id")){
+                $app_id = $this->request->param("app_id");
+                $applist = $applist->where("b.app_id",$app_id);
+            }
+            if($this->request->param("activate_state")){
+                $activate_state = $this->request->param("activate_state");
+                $applist = $applist->where("a.activate_state",$activate_state);
+            }
+            if($this->request->param("user_idfa")){
+                $user_idfa = $this->request->param("user_idfa");
+                $applist = $applist->where("a.user_idfa",$user_idfa);
+            }
             if($this->request->param("page")){
                 $page = $this->request->param("page");
             }
             
-        }
+        
         $applist = $applist->limit(($page-1)*$page_size,$page_size);
         $applist = $applist->select();
         if($applist){
@@ -149,10 +195,14 @@ class Index extends \app\common\controller
         $pageinfo["page_num"] = \ceil($page_count/$page_size);
         $userlist = Db::table('cj_user')
                            ->select();
+   
         return $this->fetch("",['applist'=>$applist,"pageinfo"=>$pageinfo,"userlist"=>$userlist]);
     }
 
     public function app_list_del(){
+        if(\json_decode(session('userinfo'),true)['user_id'] != 1){
+            return "没有权限";
+        }
         $id = $this->request->param("id");
         if(Db::table('cj_ad_click_log')->delete($id))
         {
@@ -164,6 +214,9 @@ class Index extends \app\common\controller
 
     public function app_upload()
     {
+        if(\json_decode(session('userinfo'),true)['user_id'] != 1){
+            return "没有权限";
+        }
         $page = 1;
         $page_size = 20;
         $page_count = count(Db::table('cj_ad_app')->select());
@@ -210,6 +263,9 @@ class Index extends \app\common\controller
 
     public function app_info()
     {
+        if(\json_decode(session('userinfo'),true)['user_id'] != 1){
+            return "没有权限";
+        }
         $arr = [];
            
         if($this->request->param()){
@@ -254,6 +310,9 @@ class Index extends \app\common\controller
 
     public function userlist()
     {
+        if(\json_decode(session('userinfo'),true)['user_id'] != 1){
+            return "没有权限";
+        }
         $page = 1;
         $page_size = 20;
         $page_count = count(Db::table('cj_user')->select());
@@ -262,15 +321,16 @@ class Index extends \app\common\controller
         if($this->request->param()){  
             if($this->request->param("kw")){
                 $kw = $this->request->param("kw");
-                $userlis = $userlist->where("user_account ='".$kw."' or user_name= '".$kw."'");
+                $userlist = $userlist->where("user_account","like","%$kw%")->whereOr("user_name","like","%$kw%");
             }
             if($this->request->param("page")){
                 $page = $this->request->param("page");
             }
             
         }
-        $userlis = $userlist->limit(($page-1)*20,20);
+        $userlist = $userlist->limit(($page-1)*20,20);
         $userlist = $userlist->select();
+
         if($userlist){
             $userlist = $userlist;
         }else{
@@ -285,7 +345,9 @@ class Index extends \app\common\controller
 
     public function user(){
         $arr = [];
-        
+        if(\json_decode(session('userinfo'),true)['user_id'] != 1){
+            return "没有权限";
+        }
         if($this->request->param()){
             $do = $this->request->param("do");
             
@@ -301,7 +363,8 @@ class Index extends \app\common\controller
                
             }elseif($do == 'add'){
                 $data = $this->request->post();
-
+                unset($data['do']);
+                $data['user_account'] = trim($data['user_account']);
                 if($data['user_password']){
                     $data['user_password'] = md5($data['user_password']);
                 }else{
@@ -332,5 +395,5 @@ class Index extends \app\common\controller
         session('userinfo', null);
         return $this->fetch('login');
     }
-
+    
 }
